+++ src/arp.rs
@@ -1,3 +1,5 @@
+// Listen for arp packets.
+
 use libarp::arp::ArpMessage;
 use libarp::client::ArpClient;
 use tokio::sync::mpsc;
@@ -7,7 +9,8 @@ pub struct ArpPacket {
     pub ifname: String,
 }
 
-pub async fn listen(ifname: String, arp_tx: mpsc::Sender<ArpPacket>) {
+// Listens for arp packets and sends them through arp_tx.
+pub async fn listen_loop(ifname: String, arp_tx: mpsc::Sender<ArpPacket>) {
     let mut client = ArpClient::new_with_iface_name(&ifname).unwrap();
     println!("listening on {}", ifname);
     loop {
